requirements-daemon.txt
# Minimal dependencies for headless daemon (run_daemon.py)
# Skips Flask, Perspective, Tornado — those are only needed for the web UI

requests==2.31.0
pandas==2.1.3
pyarrow>=14.0.0
websockets==12.0

# Telegram bot (with job-queue scheduler support)
python-telegram-bot[job-queue]>=21.0

# Timezone
pytz>=2023.3

# Env file loading
python-dotenv>=1.0.0

# Google Drive upload
google-api-python-client>=2.100.0
google-auth>=2.23.0
